Extract reducer to its own class
